events/serializers.py
from django.forms import widgets
from rest_framework import serializers
from .models import Event


class EventSerializer(serializers.ModelSerializer):

    class Meta:
        model = Event
        fields = ('id', 'title', 'time', 'category', 'person')
