requirements.txt
Django==1.7.6
djangorestframework==3.1.0
httpie==0.9.2
ipython==3.0.0
Pygments==2.0.2
requests==2.5.3
